Update type react-onclickoutside
