+++ definitions/npm/react-onclickoutside_v6.x.x/flow_v0.57.1-/react-onclickoutside_v6.x.x.js
@@ -1,0 +1,16 @@
+declare module 'react-onclickoutside' {
+  declare export var IGNORE_CLASS_NAME: 'ignore-react-onclickoutside';
+
+  declare export default function onClickOutsideHOC<Props: {}>(
+    WrappedComponent: React$ComponentType<Props>,
+    config?: { excludeScrollbar?: boolean }):
+      React$ComponentType<$Diff<Props, {
+        disableOnClickOutside?: () => void,
+        enableOnClickOutside?: () => void,
+        eventTypes?: Array<string>,
+        excludeScrollbar?: boolean,
+        outsideClickIgnoreClass?: string,
+        preventDefault?: boolean,
+        stopPropagation?: boolean,
+      }>>;
+}
